Add stub input and canvas
