+++ package.json
@@ -18,6 +18,7 @@
     },
     "devDependencies": {
         "@trivago/prettier-plugin-sort-imports": "^3.3.0",
+        "@types/fabric": "^4.5.12",
         "@types/wordpress__block-editor": "^7.0.0",
         "@types/wordpress__blocks": "^11.0.7",
         "@typescript-eslint/parser": "^5.38.1",
@@ -48,6 +49,8 @@
         "@wordpress/blocks": "^11.17.0",
         "@wordpress/element": "^4.16.0",
         "@wordpress/i18n": "^4.18.0",
+        "fabric": "^5.2.4",
+        "fabricjs-react": "^1.0.8",
         "swr": "^1.3.0",
         "zustand": "^4.1.1"
     }
