+++ src/components/UserInterface.tsx
@@ -11,7 +11,6 @@ import classNames from 'classnames';
 import { AnimatePresence, motion } from 'framer-motion';
 import { useModel } from '../hooks/useModel';
 import { usePrediction } from '../hooks/usePrediction';
-import { downloadImage, copyImage } from '../lib/image';
 import { useAuthStore } from '../state/auth';
 import { useGlobalState } from '../state/global';
 import { useInputsState } from '../state/inputs';
@@ -21,11 +20,11 @@ import {
     HeightInput,
     ImageLike,
     NumOutputsInput,
-    PredictionData,
     PromptInputs,
     WidthInput,
 } from '../types';
 import { InputGenerator } from './InputGenerator';
+import { MainPanel } from './MainPanel';
 import { ModelCard } from './ModelCard';
 import { GoButton } from './inputs/GoButton';
 import { FocusHelperButton } from './misc/FocusHelperButton';
@@ -174,6 +173,7 @@ export const UserInferface = ({
             : undefined;
         setPromptInputData({
             prompt: inputs?.prompt,
+            initImage: inputs?.init_image,
             width: maybeWidth
                 ? {
                       ...(maybeWidth as WidthInput),
@@ -225,7 +225,7 @@ export const UserInferface = ({
 
     return (
         <>
-            <div className="flex flex-col justify-between md:w-96 flex-shrink-0 overflow-y-scroll bg-white">
+            <div className="flex flex-col justify-between md:w-96 flex-shrink-0 overflow-y-auto bg-white">
                 <form
                     ref={focusArea}
                     className="flex flex-col gap-y-6 flex-shrink-0 p-6 pb-0 flex-grow"
@@ -269,153 +269,8 @@ export const UserInferface = ({
                     <ModelCard modelInfo={modelInfo} />
                 </AnimatePresence>
             </div>
-            <div className="bg-gray-50 flex flex-col h-full items-center overflow-y-scroll p-6 w-full relative">
-                <AnimatePresence>
-                    <MainPanel setImage={setImage} prediction={prediction} />
-                </AnimatePresence>
-            </div>
+            <MainPanel setImage={setImage} prediction={prediction} />
         </>
-    );
-};
-
-type MainPanelProps = {
-    prediction?: PredictionData;
-    setImage: (image: ImageLike) => void;
-};
-const MainPanel = ({ prediction, setImage }: MainPanelProps) => {
-    const { width, height, prompt } = useInputsState();
-    const { id, output, input } = prediction || {};
-    const { importingMessage } = useGlobalState();
-
-    if (prediction?.status === 'succeeded' && output?.length) {
-        return (
-            <div
-                className={classNames(
-                    'flex flex-col h-full w-full items-center',
-                    {
-                        'pointer-events-none': importingMessage.length > 0,
-                    },
-                )}>
-                {output?.length > 1 && (
-                    <p className="font-mono m-0 mb-4 text-center">
-                        {__(
-                            'Hover over an image for options',
-                            'stable-diffusion',
-                        )}
-                    </p>
-                )}
-                <motion.div
-                    key={prediction.id}
-                    initial={{ opacity: 0 }}
-                    animate={{ opacity: 1 }}
-                    className={classNames('gap-6 grid w-full self-start', {
-                        'lg:grid-cols-2': output?.length >= 2,
-                    })}>
-                    {prediction?.output?.map((url, i) => (
-                        <div
-                            key={url}
-                            className={classNames(
-                                'flex items-center justify-center group',
-                                {
-                                    'relative bg-gray-100':
-                                        output?.length !== 1,
-                                },
-                            )}>
-                            <div
-                                className={classNames(
-                                    'w-full bg-no-repeat bg-contain',
-                                    {
-                                        '': output?.length === 1,
-                                    },
-                                )}
-                                style={{
-                                    maxWidth: `${width}px`,
-                                    aspectRatio: `${width}/${height}`,
-                                    backgroundImage: `url(${url})`,
-                                }}>
-                                <ImageActions
-                                    id={`${id}-image-${i}`}
-                                    url={url}
-                                    caption={input?.prompt || prompt}
-                                    setImage={setImage}
-                                    forceShowHud={output?.length === 1}
-                                />
-                            </div>
-                        </div>
-                    ))}
-                </motion.div>
-            </div>
-        );
-    }
-    const imageOutput = {
-        maxWidth: `${width}px`,
-        maxHeight: 'calc(100% - 2.5rem)',
-        aspectRatio: `${width}/${height}`,
-    };
-    return (
-        <motion.div
-            key="canvas-placeholder"
-            role="button"
-            transition={{ type: 'Tween' }}
-            className="border border-gray-500 flex items-center justify-center bg-cover m-auto"
-            animate={imageOutput}
-            initial={imageOutput}>
-            <div className="w-screen" />
-        </motion.div>
-    );
-};
-
-type ActionProps = {
-    id: string;
-    url: string;
-    caption: string;
-    forceShowHud?: boolean;
-    setImage: (image: ImageLike) => void;
-};
-const ImageActions = ({
-    setImage,
-    id,
-    url,
-    caption,
-    forceShowHud,
-}: ActionProps) => {
-    const btnClass =
-        'bg-gray-900 text-white p-2 px-4 text-left outline-none focus:shadow-none focus:ring-wp focus:ring-wp-theme-500 cursor-pointer hover:bg-wp-theme-500 transition-all duration-200';
-    const { setImportingMessage } = useGlobalState();
-    const handleImport = () => {
-        if (!id) return;
-        setImage({ id, url, caption });
-    };
-    const handleDownload = async () => {
-        await downloadImage(url, `block-diffusion-${id}`);
-    };
-    const handleCopy = async () => {
-        setImportingMessage(__('Copying...', 'stable-diffusion'));
-        await copyImage(url);
-        setImportingMessage('');
-    };
-    return (
-        <div
-            className={classNames(
-                'absolute top-0 left-0 flex flex-col items-start gap-1 pt-1 group-hover:opacity-100 group-focus-within:opacity-100 transition-opacity duration-300',
-                {
-                    'lg:opacity-0': !forceShowHud,
-                },
-            )}
-            style={{
-                background:
-                    'radial-gradient(100% 65px at left top, rgb(255 255 255 / 65%) 0px, rgb(0 0 0 / 0%))',
-            }}>
-            <button className={btnClass} onClick={handleImport} type="button">
-                {__('Import into editor', 'stable-diffusion')}
-            </button>
-            <button className={btnClass} onClick={handleDownload} type="button">
-                {__('Download', 'stable-diffusion')}
-            </button>
-            <button className={btnClass} onClick={handleCopy} type="button">
-                {__('Copy', 'stable-diffusion')}
-            </button>
-        </div>
     );
 };
